+++ protractor.config.js
@@ -3,6 +3,7 @@ exports.config = {
     getPageTimeout: 30000,
     seleniumAddress: 'http://localhost:4444/wd/hub',
     specs: ['src/tests/**/*.spec.js'],
+    exclude: ['src/tests/**/TBD/*.spec.js'],
     jasmineNodeOpts: {
         defaultTimeoutInterval: 60000
     },
@@ -10,6 +11,8 @@ exports.config = {
     capabilities: {
          //'browserName': 'chrome'
          'browserName': 'phantomjs',
+         'maxInstances': 2,
+         'shardTestFiles': true,
          'phantomjs.binary.path': require('phantomjs-prebuilt').path,
          'phantomjs.cli.args': ['--webdriver-loglevel=ERROR', '--local-storage-path=/tmp/phantom_' + Math.random()]
