+++ src/tests/work-item/work-item-list/TBD/chat.spec.js
@@ -12,26 +12,24 @@
  * @author ldimaggi
  */
 
-var CommonPage = require('./page-objects/common.page'),
-  MicroservicesPage = require('./page-objects/microservices.page'),
-  testSupport = require('./testSupport');
+var CommonPage = require('../page-objects/work-item-list.page'),
+  ChatPage = require('../page-objects/chat.page'),
+  testSupport = require('../testSupport');
 
-describe('Microservices page', function () {
+describe('Chat page', function () {
   var page, browserMode;
   var until = protractor.ExpectedConditions;
   var waitTime = 30000;
 
   beforeEach(function () {
-    testSupport.setBrowserMode('phone');
+    testSupport.setBrowserMode('desktop');
     page = new CommonPage();
   });
 
   it('should contain correct page title.', function() {
-// Commented out pending resolution of: https://github.com/almighty/almighty-ui/issues/472      
-//      var theMicroservicesPage = page.clickMicroServicesMenuTab();
-//      expect(browser.getCurrentUrl()).toEqual('http://localhost:8088/intro');
-//      browser.wait(until.urlContains('intro'), waitTime, 'Failed to open microservices page');
+      var theChatpage = page.clickChatMenuTab();
+      expect(browser.getCurrentUrl()).toEqual('http://localhost:8088/chat');
+      browser.wait(until.urlContains('chat'), waitTime, 'Failed to open chat page');
   });
 
 });
-
