+++ src/tests/work-item/work-item-list/TBD/microservices.spec.js
@@ -12,11 +12,11 @@
  * @author ldimaggi
  */
 
-var CommonPage = require('./page-objects/common.page'),
-  NotificationsPage = require('./page-objects/notifications.page'),
-  testSupport = require('./testSupport');
+var CommonPage = require('../page-objects/common.page'),
+  MicroservicesPage = require('../page-objects/microservices.page'),
+  testSupport = require('../testSupport');
 
-describe('Notifications page', function () {
+describe('Microservices page', function () {
   var page, browserMode;
   var until = protractor.ExpectedConditions;
   var waitTime = 30000;
@@ -27,9 +27,11 @@ describe('Notifications page', function () {
   });
 
   it('should contain correct page title.', function() {
-      var theNotificationsPage = page.clickNotificationsMenuTab();
-      expect(browser.getCurrentUrl()).toEqual('http://localhost:8088/notifications');
-      browser.wait(until.urlContains('notifications'), waitTime, 'Failed to open notifications page');
+// Commented out pending resolution of: https://github.com/almighty/almighty-ui/issues/472      
+//      var theMicroservicesPage = page.clickMicroServicesMenuTab();
+//      expect(browser.getCurrentUrl()).toEqual('http://localhost:8088/intro');
+//      browser.wait(until.urlContains('intro'), waitTime, 'Failed to open microservices page');
   });
 
 });
+
